Move helpers into installable wisdem_interface package
