--- IEA-scaled/NREL-1.7-103/WISDEM/helpers.py
@@ -1,1 +1,0 @@
-../../../helpers.py
