--- IEA-scaled/NREL-2.5-116/WISDEM/helpers.py
@@ -1,1 +1,0 @@
-../../../helpers.py
