--- IEA-scaled/NREL-2.8-127/WISDEM/helpers.py
@@ -1,1 +1,0 @@
-../../../helpers.py
